fix(breadcrumbs): split utils for Fast Refresh, fix /admin/materials 404
- Extract ROUTE_LABELS, NON_NAVIGABLE_PATHS, buildBreadcrumbs, isDynamicSegment
  into breadcrumbs.utils.js — fixes react-refresh/only-export-components violation
  (Vite was doing full page reloads instead of HMR on Breadcrumbs.jsx edits)
- Add /admin/materials to NON_NAVIGABLE_PATHS — fixes real 404 bug where
  navigating to /admin/materials/import rendered a clickable "Materials" crumb
  that routed to a 404 page
- Update Breadcrumbs.test.jsx imports to pull utils from breadcrumbs.utils.js
- Add regression test for /admin/materials non-navigable behavior

Closes #528

Agent-Session: 2794d0a8-a9e9-4245-966f-84a5987354c5
